refactor(select): getLocale exported from configureLocalization
