+++ src/components/select/bl-select.stories.mdx
@@ -279,7 +279,7 @@ Select component can be disabled by using `disabled` attribute.
 <bl-badge icon="document">[ADR](https://github.com/Trendyol/baklava/issues/265#issuecomment-1845414216)</bl-badge>
 
 Select component can be searchable by using `search-bar` attribute.
-
+Search works with internalization support regarding to html lang attribute.
 Display a loading icon in place of the search icon with using `search-bar-loading-state` attribute, seamlessly integrating it with your API endpoint to provide a smooth user experience during data retrieval.
 
 <Canvas>
