+++ src/components/select/bl-select.ts
@@ -7,8 +7,8 @@ import { msg, localized } from "@lit/localize";
 import { FormControlMixin, requiredValidator } from "@open-wc/form-control";
 import { FormValue } from "@open-wc/form-helpers";
 import "element-internals-polyfill";
+import { getLocale } from "../../localization";
 import { event, EventDispatcher } from "../../utilities/event";
-import { getLocale } from "../../utilities/getLocale";
 import { stringBooleanConverter } from "../../utilities/string-boolean.converter";
 import "../button/bl-button";
 import BlCheckbox from "../checkbox-group/checkbox/bl-checkbox";
